sınıf geçme

+++ src/NotOrtalama/NotOrt.java
@@ -18,7 +18,7 @@ public class NotOrt {
         System.out.print("Lütfen notu giriniz: ");
         muzik = not.nextInt();
         int toplam =mat+ fizik +kimya +tarih+muzik;
-        double ortalama= toplam / 5;
+        double ortalama= (double) toplam / 5;
         boolean sonuc = ortalama > 60;
         System.out.println(sonuc ? "Sınıfı geçtiniz!" : "Sınıfta kaldınız.");
         not.close();
